[YSK-58] fix: 코드래빗 개선

+++ src/main/java/com/stockit/backend/feature/inventory/vo/InventoryQuery.java
@@ -81,9 +81,9 @@ public final class InventoryQuery {
     public String getSortColumn() { return sortColumn; }
     public String getSortDirection() { return sortDirection; }
     public LocalDate getAsOfDate() { return asOfDate; }
-    public int getOffset() { return offset(); }
+    public long getOffset() { return offset(); }
 
-    public int offset() {
-        return (page - 1) * size;
+    public long offset() {
+        return (long) (page - 1) * size;
     }
 }
